Add ThreadCard and CreateThreadDialog components: implement thread display and creation functionality, including like, report, and delete actions. Update ThreadListPage to integrate new components and handle user interactions for thread management.
