+++ apps/frontend/src/pages/Jobs/__tests__/JobDetail.test.tsx
@@ -111,7 +111,9 @@ beforeEach(() => {
   );
 
   mockSubmitMutateAsync.mockClear();
-  mockSubmitMutateAsync.mockResolvedValue(createMockApplication('1', 'PENDING'));
+  mockSubmitMutateAsync.mockResolvedValue(
+    createMockApplication('1', 'PENDING')
+  );
 });
 
 describe('JobDetail', () => {
