+++ apps/frontend/src/utils/AllTheProviders.tsx
@@ -1,25 +1,25 @@
-import { QueryClientProvider } from "@tanstack/react-query";
-import { BrowserRouter } from "react-router-dom";
-import { QueryClient } from "@tanstack/react-query";
-import AuthProvider from "../providers/AuthProvider";
+import { QueryClientProvider } from '@tanstack/react-query';
+import { BrowserRouter } from 'react-router-dom';
+import { QueryClient } from '@tanstack/react-query';
+import AuthProvider from '../providers/AuthProvider';
 
 // Create a custom render method that includes providers
 const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
-    const queryClient = new QueryClient({
-      defaultOptions: {
-        queries: {
-          retry: false,
-        },
+  const queryClient = new QueryClient({
+    defaultOptions: {
+      queries: {
+        retry: false,
       },
-    });
-  
-    return (
+    },
+  });
+
+  return (
     <QueryClientProvider client={queryClient}>
       <AuthProvider>
         <BrowserRouter>{children}</BrowserRouter>
       </AuthProvider>
     </QueryClientProvider>
-    );
+  );
 };
 
 export default AllTheProviders;
